refactor: remove enums file
