+++ src/features/game/fields/AssembleRow.ts
@@ -1,9 +1,10 @@
 import Row from "./Row";
+import { RowType } from "./RowType";
 import RoundState from "../model/RoundState";
 import HintSettings from "../model/HintSettings";
 
 import { Publisher } from "../../../shared/Observer";
-import { RowType, StageStatus } from "../enums";
+import { StageStatus } from "../model/StageStatus";
 
 import styles from "./Row.module.css";
