+++ src/features/game/fields/RowType.ts
@@ -1,0 +1,6 @@
+export const RowType = {
+  PICK: "pickArea",
+  ASSEMBLE: "assembleArea",
+} as const;
+
+export type RowType = (typeof RowType)[keyof typeof RowType];
