+++ src/features/game/model/RoundState.ts
@@ -2,7 +2,8 @@ import State from "../../../app/state/StatePublisher";
 import RoundSettings, { RoundResult } from "./RoundSettings";
 import { Observer, Publisher } from "../../../shared/Observer";
 
-import { RowType, StageStatus } from "../enums";
+import { StageStatus } from "./StageStatus";
+import { type RowType } from "../fields/RowType";
 import { type Word } from "../card/WordCard";
 import {
   generateStageWords,
