+++ src/features/game/model/StageStatus.ts
@@ -1,3 +1,4 @@
+// to avoid dependency cycles
 export const StageStatus = {
   NOT_COMPLETED: "notCompleted",
   CORRECT: "correct",
@@ -6,10 +7,3 @@ export const StageStatus = {
 } as const;
 
 export type StageStatus = (typeof StageStatus)[keyof typeof StageStatus];
-
-export const RowType = {
-  PICK: "pickArea",
-  ASSEMBLE: "assembleArea",
-} as const;
-
-export type RowType = (typeof RowType)[keyof typeof RowType];
